Group population by quintiles based on their poverty probability

Group population by quintiles based on their poverty probability

quintile_ppi(ppi, probs = c(0.2, 0.4, 0.6, 0.8, 1))

Arguments

ppi
A numeric vector of poverty probabilities or likelihoods.

probs
A numeric vector between 0 to 1 of quantile probabilities. Default values are a vector of equal probabilities (0.2, 0.4, 0.6, 0.8, 1).

Value

A numeric vector between 1 to 5 classifying data into quintiles of increasing poverty probabilities.

Examples

# ppi based on USAID extreme poverty definition using Bangladesh data from
# washdata package and PPI lookup table from ppitables package
# install.packages("washdata"); library(washdata)
ppi <- get_ppi_cohort(index = washdata::indicatorsDataBGD$ppi,
                      ppiTable = ppitables::ppiBGD2013)
quintile_ppi(ppi = ppi$extreme)

#>    [1] 3 2 3 4 4 4 5 5 2 4 1 2 2 4 4 5 2 4 3 4 3 1 4 2 2 4 4 4 2 5 2 4 5 5 4 3 4
#>   [38] 3 4 5 2 2 4 5 4 3 5 4 4 1 4 4 3 5 3 4 4 4 5 4 4 4 4 3 3 5 1 1 4 1 3 1 3 3
#>   [75] 2 2 1 2 1 1 2 1 1 1 5 2 1 1 1 1 2 2 3 3 5 1 1 2 3 2 2 1 1 3 1 2 2 1 1 1 1
#>  [112] 1 3 1 1 3 1 3 1 1 2 1 5 2 1 2 1 1 2 4 3 1 3 2 5 3 4 4 5 5 4 4 5 4 5 5 4 5
#>  [149] 5 4 4 1 4 4 4 4 1 4 1 5 2 4 5 4 5 4 5 4 4 5 4 4 5 3 3 4 4 4 3 5 3 5 4 4 4
#>  [186] 3 3 3 4 4 3 4 4 4 5 3 1 4 4 1 2 3 3 4 2 4 3 5 4 2 5 5 5 4 5 4 4 4 4 2 2 5
#>  [223] 4 4 1 1 5 3 5 5 3 2 5 4 5 4 4 4 5 3 1 2 3 5 4 2 3 3 4 4 3 2 4 3 4 3 4 3 2
#>  [260] 2 1 3 3 5 3 4 3 3 4 5 4 4 5 4 3 4 5 5 3 5 3 5 2 3 4 2 5 3 1 3 4 5 1 2 1 1
#>  [297] 1 1 2 5 2 1 1 2 5 5 4 2 4 1 4 1 4 5 2 3 1 4 1 1 1 1 1 3 5 4 3 3 1 1 5 1 5
#>  [334] 2 5 5 2 2 1 3 1 1 4 3 1 1 4 4 5 5 5 4 1 2 3 4 1 1 4 5 4 4 5 4 4 2 5 5 4 4
#>  [371] 1 4 4 3 4 4 3 5 5 5 4 4 4 3 5 5 3 2 5 3 3 3 4 5 1 3 1 4 3 4 2 2 4 4 3 4 3
#>  [408] 5 4 4 3 3 4 2 4 5 4 2 5 2 4 5 4 4 2 3 1 2 4 3 1 2 1 3 2 1 1 3 5 4 1 3 1 1
#>  [445] 1 1 3 2 1 2 4 3 1 1 1 1 2 1 2 2 1 3 1 2 1 2 2 2 2 1 5 1 4 5 4 5 5 2 3 2 1
#>  [482] 1 4 3 2 1 1 1 1 1 4 5 5 5 4 3 4 4 4 4 4 5 2 3 4 4 5 5 4 2 3 3 5 3 4 5 5 2
#>  [519] 4 5 2 2 2 2 5 5 4 3 4 5 1 3 4 4 1 5 5 2 2 4 4 3 4 5 4 5 2 4 2 3 4 3 4 5 5
#>  [556] 5 1 1 1 3 1 2 1 2 1 2 3 3 1 1 1 1 1 1 3 1 1 1 3 3 1 1 3 2 1 1 3 1 1 1 2 3
#>  [593] 1 2 4 5 2 1 1 1 2 3 1 1 1 1 1 1 2 1 4 3 3 1 1 1 2 1 2 1 1 1 5 2 3 5 4 4 4
#>  [630] 4 3 5 4 3 3 4 3 4 2 2 5 3 4 2 5 5 2 5 3 4 4 4 4 4 3 3 3 2 3 3 4 3 1 4 1 3
#>  [667] 3 3 1 3 4 4 4 2 4 4 4 2 4 4 3 3 5 3 4 4 4 3 1 1 1 1 1 1 1 2 1 3 1 1 1 1 1
#>  [704] 1 1 1 1 3 1 1 1 2 1 1 1 2 1 2 3 3 2 1 3 1 3 1 4 1 2 1 1 1 1 2 2 1 1 3 1 1
#>  [741] 1 1 1 1 2 1 1 1 1 2 1 1 1 2 1 5 3 3 5 2 3 4 2 2 4 4 4 1 3 4 4 5 4 5 4 3 3
#>  [778] 4 1 4 3 4 3 2 4 3 4 3 2 4 5 5 5 5 2 1 3 2 3 4 2 4 3 5 4 3 4 4 3 4 3 3 5 4
#>  [815] 5 5 4 2 4 5 2 1 1 1 3 1 1 1 1 1 3 4 1 1 1 1 2 3 2 1 1 2 1 2 1 4 1 1 1 2 4
#>  [852] 1 1 1 3 1 1 1 5 4 2 1 1 1 2 2 1 1 1 2 1 4 1 2 1 1 4 2 1 4 3 2 3 3 2 2 5 4
#>  [889] 3 2 3 5 5 4 4 4 4 5 5 5 4 3 2 4 4 3 3 5 4 3 5 2 4 2 4 5 4 3 4 4 3 2 4 1 3
#>  [926] 4 2 3 3 4 3 5 5 5 4 4 4 2 5 3 5 3 3 4 4 3 2 4 4 4 3 4 2 1 2 3 2 1 3 2 2 4
#>  [963] 1 3 2 2 1 2 2 2 3 4 2 4 3 4 3 3 1 3 3 1 1 1 2 1 2 2 5 3 2 3 3 4 2 3 1 1 2
#> [1000] 1 1 4 1 5 1 3 3 2 4 2 2 4 2 1 4 4 3 3 3 4 4 2 3 3 4 2 2 2 1 2 2 1 4 4 4 3
#> [1037] 4 4 4 2 2 3 3 3 4 3 3 2 5 4 3 4 4 4 4 5 4 3 4 2 3 5 4 3 4 2 3 3 1 3 3 3 3
#> [1074] 1 4 4 4 4 1 1 4 5 2 2 1 1 1 2 2 2 1 1 1 1 1 1 2 1 2 1 1 1 1 1 1 1 1 1 2 1
#> [1111] 3 1 1 1 1 1 3 2 1 1 1 3 3 2 1 2 3 1 2 3 1 2 1 1 1 1 1 3 1 1 2 2 1 3 1 1 1
#> [1148] 1 2 1 1 5 3 2 4 2 2 4 5 3 2 3 2 4 3 3 3 4 4 4 4 3 5 4 1 4 4 4 5 4 4 4 4 4
#> [1185] 4 4 3 3 4 2 3 3 4 3 4 3 3 5 5 3 3 4 3 3 2 4 5 2 3 5 5 5 5 5 5 4 3 1 1 1 2
#> [1222] 1 2 1 1 1 1 2 2 3 1 1 1 1 1 1 2 3 4 1 5 3 1 3 1 1 2 1 2 5 2 2 1 1 4 2 3 1
#> [1259] 1 1 1 2 1 1 1 2 2 2 1 2 3 1 2 1 1 1 2 1 1 4 3 1